backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
openai==1.99.5
httpx==0.25.2
PyMuPDF==1.23.8
Pillow==10.1.0
python-dotenv==1.0.0
pydantic==2.5.0
aiofiles==23.2.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
